refactor: update ombul actions
