+++ src/actions/ombul/update.ts
@@ -1,25 +1,23 @@
 'use server'
 import { updateOmbulSchema, updateObuleFileSchema } from './schema'
 import { createActionError, createPrismaActionError, createZodActionError } from '@/actions/error'
-import prisma from '@/prisma'
-import createFile from '@/server/store/createFile'
-import destroyFile from '@/server/store/destroyFile'
 import { getUser } from '@/auth/user'
 import type { UpdateOmbulSchemaType, UpdateOmbulFileSchemaType } from './schema'
 import type { ExpandedOmbul } from '@/server/ombul/Types'
 import type { ActionReturn } from '@/actions/Types'
+import { updateOmbul, updateOmbulFile } from '@/server/ombul/update'
 
 /**
- * Update an ombul
+ * A action to update an ombul
  * @param id - The id of the ombul to update
  * @param rawdata - The new data for the ombul including: name, year, issueNumber, description,
  * @returns The updated ombul
  */
-export async function updateOmbul(
+export async function updateOmbulAction(
     id: number,
     rawdata: FormData | UpdateOmbulSchemaType
 ): Promise<ActionReturn<ExpandedOmbul>> {
-    // auth route
+    // Auth route
     const { status, authorized } = await getUser({
         requiredPermissions: ['OMBUL_UPDATE']
     })
@@ -27,42 +25,23 @@ export async function updateOmbul(
         return createActionError(status)
     }
 
+    //Parse the data
     const parse = updateOmbulSchema.safeParse(rawdata)
     if (!parse.success) {
         return createZodActionError(parse)
     }
     const data = parse.data
 
-    try {
-        const ombul = await prisma.ombul.update({
-            where: {
-                id
-            },
-            data,
-            include: {
-                coverImage: {
-                    include: {
-                        image: true
-                    }
-                }
-            }
-        })
-        return {
-            success: true,
-            data: ombul
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+    return await updateOmbul(id, data)
 }
 
 /**
- * Update the ombul file (i.e. the pdf file) of an ombul
+ * A action that updates the ombul file (i.e. the pdf file) of an ombul
  * @param id - The id of the ombul to update
  * @param rawData - The new data for the new ombul file with field name 'file'
  * @returns The updated ombul
  */
-export async function updateOmbulFile(
+export async function updateOmbulFileAction(
     id: number,
     rawData: FormData | UpdateOmbulFileSchemaType
 ): Promise<ActionReturn<ExpandedOmbul>> {
@@ -80,47 +59,5 @@ export async function updateOmbulFile(
     }
     const data = parse.data
 
-    const ret = await createFile(data.ombulFile, 'ombul', ['pdf'])
-    if (!ret.success) return ret
-    const fsLocation = ret.data.fsLocation
-
-    const ombul = await prisma.ombul.findUnique({
-        where: {
-            id
-        }
-    })
-    if (!ombul) {
-        return createActionError('NOT FOUND', 'Ombul ikke funnet')
-    }
-
-    const oldFsLocation = ombul.fsLocation
-
-    try {
-        const ombulUpdated = await prisma.ombul.update({
-            where: {
-                id
-            },
-            data: {
-                fsLocation
-            },
-            include: {
-                coverImage: {
-                    include: {
-                        image: true
-                    }
-                }
-            }
-        })
-
-        //delete the old file
-        const delRet = await destroyFile('ombul', oldFsLocation)
-        if (!delRet.success) return delRet
-
-        return {
-            success: true,
-            data: ombulUpdated
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+    return await updateOmbulFile(id, data.ombulFile)
 }
