+++ src/app/ombul/[...yearAndName]/OmbulAdmin.tsx
@@ -2,7 +2,7 @@
 
 import styles from './OmbulAdmin.module.scss'
 import Form from '@/app/components/Form/Form'
-import { updateOmbul, updateOmbulFile } from '@/actions/ombul/update'
+import { updateOmbulAction, updateOmbulFileAction } from '@/actions/ombul/update'
 import { EditModeContext } from '@/context/EditMode'
 import NumberInput from '@/app/components/UI/NumberInput'
 import FileInput from '@/app/components/UI/FileInput'
@@ -36,8 +36,8 @@ export default function OmbulAdmin({
     const editCtx = useContext(EditModeContext)
     if (!editCtx?.editMode) return null
 
-    const updateOmbulAction = updateOmbul.bind(null, ombul.id)
-    const updateOmbulFileAction = updateOmbulFile.bind(null, ombul.id)
+    const updateOmbulActionBind = updateOmbulAction.bind(null, ombul.id)
+    const updateOmbulFileActionBind = updateOmbulFileAction.bind(null, ombul.id)
 
     const handleChange = async (newOmbul: ExpandedOmbul | undefined) => {
         if (!newOmbul) return
@@ -58,7 +58,7 @@ export default function OmbulAdmin({
                     canUpdate && (
                         <>
                             <Form
-                                action={updateOmbulAction}
+                                action={updateOmbulActionBind}
                                 successCallback={handleChange}
                                 submitText="Oppdater"
                             >
@@ -74,7 +74,7 @@ export default function OmbulAdmin({
                                 />
                             </Form>
                             <Form
-                                action={updateOmbulFileAction}
+                                action={updateOmbulFileActionBind}
                                 successCallback={handleChange}
                                 submitText="Oppdater fil"
                             >
