+++ src/app/ombul/[...yearAndName]/page.tsx
@@ -5,7 +5,7 @@ import { readOmbulAction } from '@/actions/ombul/read'
 import PdfDocument from '@/components/PdfDocument/PdfDocument'
 import SlideInOnView from '@/app/components/SlideInOnView/SlideInOnView'
 import EditableTextField from '@/app/components/EditableTextField/EditableTextField'
-import { updateOmbul } from '@/actions/ombul/update'
+import { updateOmbulAction } from '@/actions/ombul/update'
 import { getUser } from '@/auth/user'
 import Link from 'next/link'
 import { notFound } from 'next/navigation'
@@ -37,7 +37,7 @@ export default async function Ombul({ params }: PropTypes) {
     const canUpdate = user.permissions.includes('OMBUL_UPDATE')
     const canDestroy = user.permissions.includes('OMBUL_DESTROY')
 
-    const changeDescription = updateOmbul.bind(null, ombul.id)
+    const changeDescription = updateOmbulAction.bind(null, ombul.id)
 
     return (
         <div className={styles.wrapper}>
